Our homepage is busted. Find your terminal and SSH onto the server:
ssh -i ~/.ssh/KnpU-Tutorial.pem ubuntu@54.XXX.XX.XXX

Let's find out what the problem is. I recognize the error page as one that comes from Symfony... which means Symfony is running, and the error will live in its logs.

Check out the var/logs/prod.log file:

cd /var/www/project/current
tail vars/logs/prod.log

Ah, no surprise!

    Unknown database 'mootube'

So how do we setup the database? Well, it's up to you! This is just a one-time thing... so it doesn't need to be part of your deploy. You could go directly to MySQL and setup it up manually. That's super fine.
Creating the Database

I'm going to do it in Ansible! Add a new task... but move it above the cache tasks, in case the database is needed during cache warm up. Name it: 'Create DB if not exists'
We can use the doctrine:database:create command. So, use the command module... and I'll copy one of our other commands. Change it to doctrine:database:create then --if-not-exists so it won't explode if the database already exists.

If you run this command locally... well... we do have a database already. So it says:

    ... already exists. Skipped.

Copy that text. This last part is optional: we're going to configure this task to know when it was, or was not changed. Register a variable called create_db_output. Then, add changed_when set to not create_db_output.stdout|search() and paste that text.