When we deploy, our migrations run! Woohoo! Yep, we can just generate migrations and everything happens automatically on deploy. Making a Schema Change
Oooooooh, but there's a catch! Open src/AppBundle/Entity/Video.php. This entity has a field called image.
Ya know what? I'd rather call that poster, because it's the poster image for this video.

Because the annotation doesn't have a name option, renaming the property means that the column will be renamed in the database. And that means... drum roll... we need a migration!

But first, we also need to update a few parts of our code, like our fixtures. I'll search for image: and replace it with poster:
Then, open a template: app/Resources/views/default/index.html.twig. Search for .image. Ah, yes! Change that to .poster.

Brilliant! All we need to do now is write a migration to rename that column. Easy! Switch to your local terminal and run:

./bin/console doctrine:migrations:diff

Go check it out in app/DoctrineMigrations. Wow... it's actually perfect:

    ALTER TABLE video CHANGE image (to) poster...

Doctrine is smart enough to know that we should rename the column instead of dropping the old column and adding a new one.
Dangerous Deploy Ahead!

Great! Let's deploy! Right!? Sure... if you want to take your site down for a minute or two! Can you see the problem? If we deploy now, this migration will run... and about 1 minute later, the deploy will finish and the new code will be used. The problem is during that period. As soon as this migration executes, the image column will be gone... but the live site will still try to use it! That's a huge problem.

Nope, we need to be smarter: we need to write safe migrations. Here's the idea: only write migrations that add new things & never write migrations that remove things... unless that thing is not being used at all by the live site.