Sharing Files for Performance
There's one other reason you might want to share some paths: speed! Before now, our deploy is slow! You might not have noticed because of the magic of television: we're been fast-forwarding through the deploys! But, in real life, the Install Node dependencies task takes almost 2 minutes to run! Woh!

Why was it so slow? Because it needed to download all of the dependencies on every single deploy. But if we shared the node_modules/ directory, then yarn install would start with a full directory. It would only need to download any changes since the last deploy! This is an easy win!

Add node_modules to the shared paths